refactoring analyzer and adding file headers
