+++ helix-db/src/helixc/analyzer/methods/migration_validation.rs
@@ -1,6 +1,6 @@
 use crate::{
     helixc::{
-        analyzer::analyzer::Ctx,
+        analyzer::Ctx,
         generator::{
             migrations::{
                 GeneratedMigration, GeneratedMigrationItemMapping,
