+++ helix-db/src/helixc/mod.rs
@@ -1,3 +1,8 @@
+// Copyright 2025 HelixDB Inc.
+// SPDX-License-Identifier: AGPL-3.0
+
+//! HelixQL compiler, analyzer, and transpiler.
+
 pub mod analyzer;
 pub mod generator;
 pub mod parser;
